faft: Add retries to ec.send_command_get_output
Take an optional arg to retry the send_command_get_output command.
It is not unusual for the ec console output to get garbled, and so for
commands with no side effects, it is safe to retry them.

Use 3 retries in get_power_state().

BUG=b:159368495
TEST=firmware_WriteProtectFunc on leona

Change-Id: I8e29d990f9b767cbaec503fe34291e5a509bf7d2
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2893127
Tested-by: Jeremy Bettis <jbettis@chromium.org>
Auto-Submit: Jeremy Bettis <jbettis@chromium.org>
Commit-Queue: Greg Edelston <gredelston@google.com>
Reviewed-by: Greg Edelston <gredelston@google.com>