power_CPUFreq: Avoid races by not constantly futzing w/ min/max
The "power_CPUFreq" test goes through a funny little dance of setting
these files every time it wants to change frequencies:
- scaling_max_freq
- scaling_min_freq
- scaling_setspeed

It does them in a different order depending on whether it's going _up_
a frequence or down. The need to change the order makes some sense
since I can believe that the kernel might not let you request a min
that's higher than max or a max that's lower than min. However, the
dance itself is a bit pointless (or worse).

Specifically, the dance is pointless because all we really need to do
is use "scaling_setspeed" to set the frequency. After all, we're
already requiring the existence of 'userspace' governor and that's all
ya gotta do.

In fact, the dance is worse than pointless. From looking at the kernel
code (at least kernel 5.4), it appears that when you change the "min"
or "max" value that it doesn't take effect instantly. You can see that
the functions cpufreq_notifier_min() cpufreq_notifier_max() are
implemented with a simple schedule_work(). If that work doesn't get
scheduled in time then we could actually be using old constraints and
make the test fail!

It appears to have always been doing this funny little dance since the
test was first written in <https://codereview.chromium.org/603034>
(yes, it predates the Chrome OS team's use of gerrit). If there was
ever any reason to set the min and max for every frequency change then
I'm sure it's long gone.

Just to be safe, though, we'll still set the min/max one time right at
the start of the test. If anything else in the system had messed with
them then it will at least get those constraints out of the
way. Presumably nothing in the system is _actively_ messing with them
during the test since we really can't do anything about that anyway.

BUG=b:173229974
TEST=test_that ${DUT} power_CPUFreq --fast --iterations=100

Change-Id: I9f5e2f0d72765dc1f409f2110eb1dfdba15b0699
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2904712
Tested-by: Douglas Anderson <dianders@chromium.org>
Reviewed-by: Stephen Boyd <swboyd@chromium.org>
Commit-Queue: Douglas Anderson <dianders@chromium.org>