bluetooth: Refactor Bluetooth's native facades
This refactors BluetoothFacadeNative and creates two new classes:
BluezFacadeNative and FlossFacadeNative. Both inherit from
BluetoothBaseFacadeNative which will be used for common functionality.

The new Floss facade will be used to interact with the Floss bluetooth
daemon. Control files are expected to pass the Floss parameter when
tests are expected to work on the Floss daemon.

BUG=b:179715260
TEST=Ran bluetooth_AdapterSAHealth

Change-Id: I6e3f3a7872ccdb5c4ee9fe4049a83831434ebd7c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2913349
Tested-by: Abhishek Pandit-Subedi <abhishekpandit@chromium.org>
Reviewed-by: Shijin Abraham <shijinabraham@google.com>
Commit-Queue: Abhishek Pandit-Subedi <abhishekpandit@chromium.org>