hardware_StorageQual: Make bug_id and part_id optional keyvals
The bug_id and part_id keyvals should not be the hard requirement to the
tests. This prevents people from running the subtests directly through
test_that for debugging purpose.

This CL aligns with CL:1249615 and should not introduce any functional
change to the Moblab suite test.

BUG=b:171674425
TEST=local test_that without Moblab setup:
     test_that $IP hardware_StorageStress_retention_soak_0

Change-Id: Icfe9bf8ceb9ba759aa1901cc3d24e09c6f29e87e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2928696
Reviewed-by: Alexis Savery <asavery@chromium.org>
Commit-Queue: Fei Shao <fshao@chromium.org>
Tested-by: Fei Shao <fshao@chromium.org>