Bluetooth: add set_target_devices to AdvMon
If DUT is in a noisy environment or in chamber there is some other peers
advertising, it is possible that AdvMon receives DeviceFound or
DeviceLost event from peer that is not used in a test, which could fail
the test.

This patch adds a method to set target devices, and makes AdvMon only
cares about events from these target devices.

BUG=b:188722516
TEST=None

Change-Id: I203f9c2109d57fab067e392967c7de238213bb1f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2929933
Tested-by: Yun-Hao Chung <howardchung@chromium.org>
Reviewed-by: Manish Mandlik <mmandlik@chromium.org>
Reviewed-by: Archie Pusaka <apusaka@chromium.org>
Commit-Queue: Yun-Hao Chung <howardchung@chromium.org>