graphics_parallel_dEQP: Format single test failures in a single line.
A pile of test names with no newlines is illegible in stainless and in
the CLI, but Stainless and the sheriff mails only show and group
failures by first line of the failure.  Today's sheriff email had 3/6
of the "failed N dEQP tests" cases with N=1, so presenting that single
case there will be helpful in triage.

BUG=b:180147369
TEST=test_that trogdor graphics_parallel_dEQP.gles

Change-Id: I98e83042008f525ce50741a116224511630cc4c1
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2930365
Auto-Submit: Emma Anholt <anholt@google.com>
Tested-by: Emma Anholt <anholt@google.com>
Reviewed-by: Po-Hsien Wang <pwang@chromium.org>
Commit-Queue: Emma Anholt <anholt@google.com>