cellular: Remove backchannel from cellular tests
Backchannel has been the source of flakiness on the cellular tests for a
long time. The only reason to use it was to verify that the connection
to a url was done through the cellular connection and not ethernet. The
same can be accomplished with the ping command specifying the interface
to be used.

BUG=b:190229823
TEST=test_that lazor --fast suite:cellular_ota
TEST=test_that jinlon --fast suite:cellular_ota

Change-Id: I46f728dc067f756ef2ed394e866391451e31bd25
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2942210
Auto-Submit: Andrew Lassalle <andrewlassalle@chromium.org>
Tested-by: Andrew Lassalle <andrewlassalle@chromium.org>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Commit-Queue: Andrew Lassalle <andrewlassalle@chromium.org>