firmware_PDVbusRequest: select volt/curr under max_charging_power
When we limit the device max_charging_power, we need to skip
to select the power which over the max_charging_power.
For example, if we set max_charging_power = 45w,
the voltage = 20v, current = 3a should be skipped because
the max charging power is 60w.

BUG=b:186383390
TEST=test_that firmware_PDVbusRequest*

Change-Id: I99c1d8bba641360c8b45936e6c700f48e19c4e24
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2951186
Reviewed-by: Wai-Hong Tam <waihong@google.com>
Tested-by: Zhuohao Lee <zhuohao@chromium.org>
Commit-Queue: Zhuohao Lee <zhuohao@chromium.org>