power_suspend: Increase default suspend delay to 15 seconds.
It appears that chrome routinely takes >10 seconds to respond to SuspendImminent,

  powerd.LATEST
  2021-06-10T16:09:02 INFO powerd: [suspender.cc(495)] Starting request 95289346
  ...
  2021-06-10T16:09:13 INFO powerd: [suspend_delay_controller.cc(103)] Got
    notification that delay 95289353 (chrome) is ready for suspend request
    95289346 from :1.257

As this delay is part of the suspend_delay accounting it makes many
platforms fail with 'Confidence check failed: did not try to suspend'
since the actual kernel suspension doesn't occur till after that.

To fix in the short-term, doubling the suspend delay to 15.

BUG=b:184191321
TEST=power_UiResume passes reliably on hatch.

Change-Id: I3539caac15e19ed9700d1072d56145a0e237260d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2952945
Tested-by: Todd Broch <tbroch@chromium.org>
Reviewed-by: Puthikorn Voravootivat <puthik@chromium.org>
Commit-Queue: Todd Broch <tbroch@chromium.org>