wait_up: Log the moment of SSH success
FAFT tests commonly fail at host.wait_up(). This might be because of a
failure to connect to SSH. But it's also possible that the SSH
connection is live, but certain wait_up processes are not started.

This CL adds a bit of debug-level logging to distinguish between those
two scenarios.

BUG=None
TEST=empty FirmwareTest, check debug logs

Change-Id: Idf7db8b422ab6e77cc732573182069fedc6936ca
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2959202
Tested-by: Greg Edelston <gredelston@google.com>
Auto-Submit: Greg Edelston <gredelston@google.com>
Reviewed-by: Jeremy Bettis <jbettis@chromium.org>
Commit-Queue: Jeremy Bettis <jbettis@chromium.org>