autotest: Delete restricted_user and TKO CGI
As part of shadow_config audit, deleting the restricted_user config.
This has to do with the user management in AFE, where some users can
have restricted permissions.  Needless to say, the entire concept is
obsolete, so delete it.

The only reference to restricted_user is in the TKO CGI pages.  Since
the TKO results web UI is also obsolete, take the opportunity to
delete all of the CGI files, since it's low risk.

BUG=b:191304437
TEST=None

Change-Id: Ife2f025ceaaa0bce7eba57fd25e1b575904a719e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2977436
Tested-by: Allen Li <ayatane@chromium.org>
Commit-Queue: Allen Li <ayatane@chromium.org>
Reviewed-by: Derek Beckett <dbeckett@chromium.org>
Reviewed-by: Congbin Guo <guocb@chromium.org>