faft: Tweak the instructions a little
You shouldn't need to build_packages to run FAFT, so put that as an
alternative, not as the default.

Tast gets out of date, and update_chroot is the way to fix it, so add
that to the instructions also.

BUG=b:191391921
TEST=N/A

Change-Id: I736b08b0538fbf7f36468f1a1936d33b0320961d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2988482
Commit-Queue: Jeremy Bettis <jbettis@chromium.org>
Tested-by: Jeremy Bettis <jbettis@chromium.org>
Auto-Submit: Jeremy Bettis <jbettis@chromium.org>
Reviewed-by: Kevin Shelton <kmshelton@chromium.org>
Reviewed-by: Derek Beckett <dbeckett@chromium.org>