bluetooth: If syslog file is missing, log error
If a system is missing the syslog file (/var/log/messages), LogManager
will raise an exception during __init__ and this causes xmlrpc to fail
to load. Rather than fail outright in this scenario, we can just use
logger to log a message to syslog, which should create the syslog file.
If there is still a problem with the syslog file getting created, it
would get caught when LogManager attempts to parse the logs and would
fail a specific test rather than outright stop execution.

BUG=b:192261342
TEST=Ran bluetooth_SAHealth.sa_basic_test

Change-Id: I18361e8ddb1f600618d424c3c5433e2e7aefaf46
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2993578
Tested-by: Abhishek Pandit-Subedi <abhishekpandit@chromium.org>
Reviewed-by: Seewai Fu <seewaifu@google.com>
Commit-Queue: Abhishek Pandit-Subedi <abhishekpandit@chromium.org>