cheets_CTS_R: Exclude an infite-loop test.
The excluded test causes an infinite loop and blocks
all the subsequent tests to run due to a bug in 11r4.
Until the fixed version (11r5) is delivered, we skip
running the problematic test case.

The change is intentionally NOT built into the control
file generator so that we don't forget to drop the
workaround when we generate r5 control files.

BUG=b:191653140
TEST=cheets_CTS_R.arm.CtsAppSecurityHostTestCases
TEST=cheets_CTS_R.11_r4.x86.CtsApp

Change-Id: I1d210a38ce5f2fc65932f9338ad40dbeebcc2f19
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2995320
Tested-by: Kazuhiro Inaba <kinaba@chromium.org>
Auto-Submit: Kazuhiro Inaba <kinaba@chromium.org>
Reviewed-by: Jiyoun Ha <jiyounha@chromium.org>