firmware_ECChargingState: Exempt TERMINATE_CHARGE alarm
TERMINATE_CHARGE_ALARM seems to represent a request to temporarily
terminate charging. From the smart battery standard [0]:

"""
TERMINATE_CHARGE_ALARM bit is set when charging should be stopped but
the Smart Battery may
not yet be in a Fully Charged state. Charging is effectively
‘suspended,’ usually temporarily. Charging may
resume when the Smart Battery detects that its charging parameters are
back in allowable ranges and
ChargingVoltage() and ChargingCurrent() values are both returned to
non-zero values. This bit is cleared
when the Smart Battery detects that it is no longer being charged.
"""

This bit may get set briefly when charging hits 100%. Exempt it from the
battery alarm detection in the FAFT test if the FULL bit is also set.
There are other more "alarming" circumstances where we'd want to catch
this bit going high, like if charging parameters were invalid. So we
only exempt this alarm if the FULL bit is also set.

The other alarms are still valid, and are not expected to fire during the
test.

[0] http://sbs-forum.org/specs/sbdat110.pdf

BUG=b:192424449
TEST=Run firmware_ECChargingState

Change-Id: I3d23d7b3283313f6fb8b9968034ea64d26865cf6
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/3001765
Tested-by: Evan Green <evgreen@chromium.org>
Reviewed-by: Aseda Aboagye <aaboagye@chromium.org>
Commit-Queue: Evan Green <evgreen@chromium.org>