servo: Add support for containerized servod
Moblab and Satlab are moving to containerized servod, this is a
first step towards that.

It emulates the localhost (but the hostname can not be localhost
and each servo will have its own hostname)

Disabing repair for now that do not work, if proof of concept goes
well we can come back and figure out how to implement the disabled
checks for containerized servod.

BUG=b:192569924
BUG=b:178502297
TEST=hacked this code into a build and ran it on moblab.

Change-Id: Ic756252f16e7ff18b699e9ac1e1c7f6ef34459e7
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/3000854
Reviewed-by: Keith Haddow <haddowk@chromium.org>
Commit-Queue: Anh Le <anhdle@chromium.org>
Tested-by: Anh Le <anhdle@chromium.org>
Auto-Submit: Anh Le <anhdle@chromium.org>