cellular: Ignore errors during clean up
After a test fails, it restores the test environment to a good state.
If restoring the state fails, the error thrown at that point should
not be reported as the reason for the test failure if a previous error
exist, because it overrides the real root cause of the failure.

BUG=None
TEST=Forced different failures and verified the logs.

Change-Id: Ic55eb64400c17b465543b21ca35817cb97b0d16a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/3033333
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Reviewed-by: Madhav . <madhavadas@google.com>
Commit-Queue: Andrew Lassalle <andrewlassalle@chromium.org>
Tested-by: Andrew Lassalle <andrewlassalle@chromium.org>