VirtualEthernetPair: Optional network namespace
Interface part of the veth pair can be moved to another network
namespace.  This allows better separation for tests that use it for
configuration of DUT.  For example DHCP server can be run on interface
and with this option it will not be visible in DUT so it will behave
as if it was an external server (no routing changes in DUT).

BUG=None
TEST=manual:
* create virtual ethernet pair with 'interface_ns' arg
* run 'setup()' on it
* check if this namespace has been created and has interface in it

Change-Id: Iece176d2ed4a8e4fd242adb88403cb827566d490
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/3043938
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Tested-by: Hugo Benichi <hugobenichi@google.com>
Commit-Queue: Hugo Benichi <hugobenichi@google.com>