bluetooth: Don't use incomplete UUIDs in ServiceAllowlist tests
After crrev/c/3033704, we no longer check UUIDs of all mandatory
services of auto-connect profiles for IsBlockedByPoliy. With this
change, an incomplete UUID allowed list might lead a not-allowed device
to indicate not blocked.

The reasons why we modify the test is because:
It is clients' duty to set service allowed list properly. It doesn't
make sense to allow HoG but disallow DeviceInfo.

BUG=b:194245291
TEST=run bluetooth_AdapterEPHealth on hatch

Change-Id: I86ab09310fbe4cbda35614b9cc8c83333772fa1e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/3043435
Tested-by: Yun-Hao Chung <howardchung@chromium.org>
Reviewed-by: Shyh-In Hwang <josephsih@chromium.org>
Commit-Queue: Yun-Hao Chung <howardchung@chromium.org>