dhcp: Fix the DHCP renew test
It turns out that version 6.8.2 of dhcpcd used in ChromeOS is not
compliant with DHCP standard when it comes to lease renewal.

The problem is that in "passive mode" dhcpcd is skipping internal
configuration and when the renew (T1) time comes the REQUEST message is
not sent according to the DHCP standard which requires that:
- it should be an unicast packet sent to the DHCP server
- it MUST NOT have "server ID" and "requested IP" DHCP options
- it MUST have client IP filled in 'ciaddr'
None of these points is satisfied.

In upgraded dhcpc (see CL:3199793) "passive mode" has been reworked and
now the internal configuration is performed - only the changes to the
system configuration are left for shill.  With this change the renew
process is now compliant and this commit fixes the problems with DHCP
renew test itself, which are:
- dhcpcd talks with mock DHCP server over virtual ethernet pair but the
  sever side is visible in DUT and it affects routing and delivery of
  unicast packets to the server
- mock DHCP server is accepting only broadcast addresses
- renew test was expecting wrong format of REQUEST message being sent
  during RENEW and REBIND phase.

BUG=b:172213156
TEST=run network_DhcpRenew test with upgraded dhcpcd

Cq-Depend: chromium:3199793
Change-Id: Idcd657d7852af39d8893f106775c1f27c372fa3e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/3168773
Tested-by: Andrzej Ostruszka <andrzejo@google.com>
Commit-Queue: Andrzej Ostruszka <andrzejo@google.com>
Reviewed-by: Billy Zhao <billyzhao@chromium.org>