autoupdate_Basic: Add login before and after the update.
There have been serveral post mortems lately related to logging in on
stable channel.

Some of these could have been caught by a test that logs in with the same user
account before and after an update between milestones. We have tests that do
logins already (paygen_au_stable) but it defaults to clearing all
existing users before login. We need a test that tries to login with an
existing user.

This CL adds that functionality to autoupdate_Basic.m2n.full that runs in a
blocking suite on the release builders and also runs in the CQ.

Tested the following scenarios with this CL and got results:
Update from 13904.77.0 to 13904.94.0: Failing to login after update.
Update from 13904.77.0 to 13904.97.0: Able to login after update.

BUG=b:194095357
BUG=b:192393289
TEST=See above

Change-Id: I709a6204b82dc2227e53c973614b136289f1622b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/3044353
Tested-by: David Haddock <dhaddock@chromium.org>
Reviewed-by: Jae Hoon Kim <kimjae@chromium.org>
Commit-Queue: David Haddock <dhaddock@chromium.org>