autotest: Allow tests to specify interfaces for Netperf tests
Currently, the classes NetperfRunner and NetperfSession assume that the
caller wants to use the WiFi interfaces of all devices used in Netperf
testing. The change updates the classes to allow the caller to specify
an interface that it would like to use on the devices under test. If no
interface is specified, the test will attempt to use the WiFi interface
by default, which should prevent this change from breaking other tests.

BUG=b:172211699
TEST=test_that --fast $DUT network_WiFi_Perf.vht80
Results are not affected by this change.

Change-Id: I0aeaffc5df129bb2ef342848c4b4a718efe256a8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/3064787
Reviewed-by: Nicolas Norvez <norvez@chromium.org>
Reviewed-by: Arowa Suliman <arowa@chromium.org>
Tested-by: Kevin Lund <kglund@google.com>
Commit-Queue: Kevin Lund <kglund@google.com>