autotest: Three way Perf setup
Change the WiFi_Perf tests to ustilize the pcap device as a Netperf
endpoint in throughput testing. This setup utilizes some hard-coded
values for ethernet interface names and ip addresses for the LAN ports
on the pcap and router Gale devices. This is necessary because the
router device is not configured with the daemons to handle the
assignment automatically. This also means that some ip routes need to be
set up manually on the devices to allow traffic to move as needed.

Specific arguments for the LAN interface device names and ip addresses
may be provided at the command line. If not provided, default values
will be used instead.

The throughput results after making this change are improved as
expected, but not completely as high as expected. There is still a
bottleneck on the router CPU in some cases. In other cases, we are
measuring the total transmitted data from the source device, rather than
measuring the received data on the target device. This results in some
test cases that have artificially inflated throughput values. These
problems could all be resolved by transition from Netperf to Iperf.

This implementation uses IpConfigContextManager to make ip config
changes on the test devices. This ensures that the ip config changes are
not persisted if the test fails or is otherwise interrupted. To test
this, I exited the test at various stages and checked that the ip
configs were correcty undone.

BUG=b:172211699
TEST=test_that $DUT suite:wifi_perf
test_that --fast $DUT network_WiFi_Perf.vht80 --args='router_lan_ip_addr="192.168.1.49"'

Change-Id: I940d37e8b59379d9bdc03d62390b00a7f718cc79
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/3108426
Commit-Queue: Kevin Lund <kglund@google.com>
Tested-by: Kevin Lund <kglund@google.com>
Reviewed-by: Arowa Suliman <arowa@chromium.org>