bluetooth: capture the ResetStack exception during test initialization
This patch fixed the exception capturing logic for device.ResetStack()
during test initialization.

There was existing code that tried to capture the chameleond reset
exception that was triggered during early test initialization.
However, the exception capturing logic seemed not working all the
time. This issue might be related with the recent underlying infra
change due to python3 migration. Hence, the exception capturing logic
should be fixed to contain the string that is observed when running
with python3. To be compatible with python2 which is still being
uses in the lab, the original logic is reserved as well.

BUG=b:196462735
TEST=Run bluetooth tests with both python2 and python3. The tests
should pass in both cases.

Change-Id: I1cd75482a8a6b572ca279f5257beb4e021fba38d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/3070859
Reviewed-by: Michael Sun <michaelfsun@google.com>
Reviewed-by: Archie Pusaka <apusaka@chromium.org>
Tested-by: Shyh-In Hwang <josephsih@chromium.org>
Commit-Queue: Shyh-In Hwang <josephsih@chromium.org>