[autotest]: Disable cellular test - cellular_ModemControl
cellular_ModemControl test being disabled as modemmanager actions
are not expected to reflect in shill in the way it is implemented in
recent.this does not reflect any user scenario. shill actions are
covered in tast test. modemmanager actions will be covered as part
of b:181016882

BUG=b:192296455
TEST=None

Change-Id: Ic0560d30416b33f0b24f999640278ccaca9b8365
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/3072322
Tested-by: Srikanth Oruganti <srikanthkumar@google.com>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Reviewed-by: Steven Bennetts <stevenjb@chromium.org>
Reviewed-by: Srikanth Oruganti <srikanthkumar@google.com>
Commit-Queue: Srikanth Oruganti <srikanthkumar@google.com>