bluetooth: fix device attribute condition
The chameleon device is special in that hasattr(device, xxx) will
evaluate to the _Method class if xxx does not physically exist. Hence,
isinstance(device.rssi, int) instead of hasattr(device, 'rssi') is
used as the condition to determine if the rssi attribute exists.
Refer to class _Method in client/cros/chameleon/chameleon.py
for more details.

BUG=b:195863578
TEST=The bluetooth_AdapterAdvMonitor.advmon_pattern_filter_tests
test should pass with or without "rssi_check=false" in the --args
of test_that.

Change-Id: Id2c1ca6420f8c0fd9667039455375d74471b3a63
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/3079861
Tested-by: Shyh-In Hwang <josephsih@chromium.org>
Commit-Queue: Shyh-In Hwang <josephsih@chromium.org>
Reviewed-by: Yun-Hao Chung <howardchung@chromium.org>
Reviewed-by: Archie Pusaka <apusaka@chromium.org>