autotest: Add IperfRunner and IperfSession
This change adds an Iperf runner and Iperf session manager to the
auotest framework, which can be used to run performance tests using the
Iperf tool.

The implementation for these wrappers is largely based on the
implementation of NetperfSession and NetperfRunner. Some changes have
been made, which are explained in the following design doc:
go/cros-wifi-iperf-design.

BUG=b:195574472
TEST=test_that $DUT suite:wifi_perf

Change-Id: I97f96a90e10bad9afa7a6e92e12ab1deae31fdbb
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/3133629
Tested-by: Kevin Lund <kglund@google.com>
Commit-Queue: Kevin Lund <kglund@google.com>
Reviewed-by: Abhishek Kumar <kuabhs@chromium.org>
Reviewed-by: Billy Zhao <billyzhao@chromium.org>