autotest: Use tool-ignorant test types for perf tests
Currently, the performance testing stack uses a lot of Netperf specific
concepts. Specifically, individual tests are defined as NetperfConfig
TestTypes, which are an internal concept to the Netperf performance
testing tool. This change replaces that logic with a tool-agnostic set
of test type definitions which describe the tests in the verbiage which
is generally used by the WiFi team, and which is not specific to any
given performance testing tool.

This change adds a new class PerfTestManager, which adds a new layer of
abstraction for Performance tests. The PerfTestManager is used to create
the required objects to run perf tests, and means that the test itself
does not need to be aware of tool-specific details.

This change is laying the groundwork to
be used when Netperf is replaced by Iperf in the testing suite.

This change only modifies the wifi_perf suite. The coex testing suite
and attenuated perf suite are untouched by this change.

BUG=b:195574472
TEST=test_that --fast $DUT network_WiFi_Perf.vht80

Change-Id: I1c388f54d9895f7867349a2e9b04ff5e983630f5
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/3088131
Tested-by: Kevin Lund <kglund@google.com>
Commit-Queue: Kevin Lund <kglund@google.com>
Reviewed-by: Matthew Wang <matthewmwang@chromium.org>
Reviewed-by: Abhishek Kumar <kuabhs@chromium.org>