autotest[PY3] - migrating dynamic_suite unittest to py3
This *will* lose a bit of coverage due to the differences of mox/mock.
I've tried to keep it as covered as possible, but since this file is
going to be refactored in a few weeks/months for tauto, I really didn't
want to spend another few days/week digging into mox.

BUG=b:187788514
TEST=this in py2/3, dummy_Pass

Change-Id: Idb1f44dd77a275b4f2f22ec904cece0d5fc8b6ce
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/3075251
Commit-Queue: Derek Beckett <dbeckett@chromium.org>
Tested-by: Derek Beckett <dbeckett@chromium.org>
Reviewed-by: C Shapiro <shapiroc@chromium.org>