autotest: start checking TCP TX throughput over 100 mbps again
Due to a bottleneck in the testbad, we had to stop verifying any
performance results for the TCP TX configuration that were over 100
mbps. Since the testbeds have been converted to a three-way performance
testing setup, we should be able to achieve 100 mbps in this
configuration again, so remove the logic that skipped verification in
these cases.

BUG=b:172211699
TEST=test_that $DUT network_WiFi_Perf.vht80
Shows throughput 375.29 mbps for TCP TX

Change-Id: I8c82ff88fb9da5face8ea439764e872278f0f724
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/3088313
Tested-by: Kevin Lund <kglund@google.com>
Commit-Queue: Kevin Lund <kglund@google.com>
Reviewed-by: Nicolas Norvez <norvez@chromium.org>
Reviewed-by: Arowa Suliman <arowa@chromium.org>