autotest: Fix test utils class object mocking
This is a follow-up CL to crrev.com/c/3075044. The change to `make_new`
in test_utils/mock.py in that CL was a workaround to get tests passing
but is not correct in the general case. Specifically, the changes in
that CL would *not* work if a test tried to mock a NamedTuple, a
subclass of `str` or `float`, or any other immutable data type.

The original code was *also* not correct. The manner in which it
forwards arguments to the mocked object's base class has raised a
warning since Python 2.6, but it only became a hard error in Python 3.3.
This hard error was what the previous CL was trying to work around as
part of the effort to migrate this project to Python 3.

So what was wrong about the original code and why is this CL the right
fix? To understand that, we have to understand what the code was
originally trying to do and what exactly is the source of the error
raised in Python 3. Trying to run the unaltered unit tests under Python
3 fails with the cryptic error `TypeError: object() takes no
parameters`. This error is raised from within the implementation of the
Python interpreter itself [1].

The code in question is in the method `create_mock_class_obj`. The
intent of this method is to create a mock object that mimics the methods
of a class. But rather than creating a mock object that pretends to be
an *instance* of a class this mock is supposed to replace the *entire
class type definition*, which means that it needs to be able to emulate
& record class methods including the class constructor. Normally,
calling a class (e.g. `MyClass(1, 2, "foo")`) would create and
initialize an instance of MyClass, and the intent of these mock class
objects is to be able to record those constructor calls so that tests
can assert and expect that certain constructor calls were made.

So how are we supposed to dynamically define a mock class that
1. is a subtype of the original class (so that `isinstance` and such
   still work correctly)
2. captures constructor calls correctly, and
3. still has some way to instantiate the original class so that the mock
   can inspect an actual instance to figure out what mock function stubs
   need to be generated?

The existing code accomplished this by overriding the __new__ method in
the mock class object. __new__ is a magic method that controls object
creation *before* that object is passed to the __init__ method of the
object [2]. Typically this magic method is used by immutable types such
as tuples to construct their values, as by the time the object would be
passed to __init__ it is already "frozen" and not permitted to be
modified by Python code. (This is why the change in crrev.com/c/3075044
is insufficient: not passing the arguments to __new__ means that
immutable types won't be constructed correctly.)

Calling a class, e.g. making a class instance via `foo = MyClass("abc",
2), has the default behavior of calling the class' __new__ method,
checking to see if the value returned from that method is an instance of
the class, and passing that value to the __init__ method of that class
if that value a proper instance. But it *is* legal behavior for __new__
to return *anything*: If __new__ returns something that doesn't look
like the requested type the constructor just discards it and returns
nothing [2]. This is the approach taken by the original code: Override
__new__ to record the object creation attempt and then return nothing.

However, the original approach was broken because it sidestepped its own
overridden __new__ method via `super` in the `make_new` method. If
nothing else in the class' inheritance hierarchy defines a custom
__new__ method the `super().__new__` call delegates all the way up the
type hierarchy to `object.__new__`, which is where the hard error seen
only in Python 3 is coming from [1]. The exact rules for when the
TypeError is raised are documented in [3], but in summary if a type
defines both a custom __init__ and a custom __new__, then that type
delegating a call to `object.__new__` and passing any arguments is
assumed to be a mistake and a TypeError is raised. A custom __init__
method is required in our code though in order to avoid mock class
instantiation from causing any side effects that may occur in the real
object's __init__, so hooking __new__ to catch object creation is not a
viable approach due to these constraints.

In Python, calling an object e.g. `foo()` delegates to the __call__
magic method of that object. Usually we think of objects as being
instance objects, but in Python class definitions themselves are also
objects! The type of a class definition is `type`, so calling a class
definition ends up delegating to the __call__ method of `type`. Since our
original goal was to capture class constructor invocations, why not
change what a call to a class does altogether rather than trying to mess
with low level object creation?

The typical way you override a type's behavior is by creating a subclass
of that type and defining custom methods with your changed behavior,
which brings us to the solution in this CL: A metaclass is just a
subtype of `type` that provides alternative definitions for the
fundamental behaviors of a class. We define a simple metaclass that
overrides __call__ to record the invocation and return nothing, and then
set this metaclass as the base type of our dynamically created mock
class object. This allows us to delete the previous custom __new__
definition, which allows us to correctly pass arguments to the
underlying class' __new__ without hitting the TypeError in
`object.__new__`.

[1] https://github.com/python/cpython/blob/v3.6.12/Objects/typeobject.c#L3512
[2] https://docs.python.org/3/reference/datamodel.html#object.__new__
[3] https://github.com/python/cpython/blob/v3.6.12/Objects/typeobject.c#L3441-L3481

BUG=b:187788514
TEST=all python2 unittests pass
TEST=all passing python3 unittests still pass

Change-Id: I9a7568e8e5e631778133a531d1be60a3db6eadc4
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/3091730
Tested-by: Chris McDonald <cjmcdonald@chromium.org>
Reviewed-by: Greg Edelston <gredelston@google.com>
Reviewed-by: Derek Beckett <dbeckett@chromium.org>
Commit-Queue: Chris McDonald <cjmcdonald@chromium.org>