autotest: Add IpConfigContextManager
Add a context manager to the Autotest tools whih allows tests to make
some changes to device ip configurations using the `ip` command. The
following capabilities are supported:
* Bring an interface up on a device.
* Assign an ip address to an interface on a device.
* Specify an ip route on the device.

This context manager tracks all the changes made to the device, and
automatically reverses all of the impacts of the changes, such as
bringing down interfaces and deleting ip addresses and routes.

BUG=b:197104627
TEST=test_that $DUT network_WiFi_Perf.ht20

Change-Id: I3ba5533e305c5ff5726e7923a3d63ea968ffae8c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/3108425
Commit-Queue: Kevin Lund <kglund@google.com>
Tested-by: Kevin Lund <kglund@google.com>
Reviewed-by: Arowa Suliman <arowa@chromium.org>