firmware_ECCbiEeprom: Update for decoupled CBI WP
This commit updates the firmware_ECCbiEeprom test to account for
platforms where the CBI EEPROM WP may be decoupled.  On those systems,
the EC has a gpio `EC_CBI_WP` that will assert according to the result
of `system_is_locked()` when the EC boots.  This signal is also latched
when it is asserted and the latch will only clear when EC_RST_ODL is
asserted.

This commit checks to see if the board has this GPIO and will assert
EC_RST_ODL after changing the system WP in order for the latch to be
reset.  Additionally, SW WP is asserted when HW WP is asserted since
that's the criteria that `system_is_locked()` will look for.

BUG=b:195077687
TEST=`cros-workon-keeby start autotest; emerge-keeby autotest; test_that
-b keeby $DUT_IP firmware_ECCbiEeprom` and verify that it works as
expected on lalala.

Change-Id: I15192c087bb508284371dff7749de1c311e7716a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/3132311
Commit-Queue: Aseda Aboagye <aaboagye@chromium.org>
Tested-by: Aseda Aboagye <aaboagye@chromium.org>
Reviewed-by: Jeremy Bettis <jbettis@chromium.org>