autotest: Move FirewallOpen logic to LinuxSystem
The function firewall_open is currently defined within the WiFiClient
class, which defines a device with WiFi connectivity. WiFiClient is a
subclass of LinuxSystem, which can be use to define any Linux device to
be used in testing. The firewall_open function is used to open the
firewall rules on a device to allow another device to establish a
connection. There is nothing about the function logic that only applies
to a WiFi enabled device, and further changes will require the firewall
rules to be opened on other Linux devices, so this change moves the
function into the LinuxSystem superclass. This change also moves the
firewall cleanup logic into the superclass as well.

BUG=b:195574472
TEST=test_that $DUT suite:wifi_perf

Change-Id: I232d6f746654762d86829fb103a02765e126be99
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/3133628
Tested-by: Kevin Lund <kglund@google.com>
Commit-Queue: Kevin Lund <kglund@google.com>
Reviewed-by: Gregory NISBET <gregorynisbet@google.com>
Reviewed-by: Otabek Kasimov <otabek@google.com>
Reviewed-by: Billy Zhao <billyzhao@chromium.org>